Firestoreへのアクセス方法をリファクタリングした
