+++ src/features/auth/auth.ts
@@ -12,7 +12,7 @@ import {
   verifyBeforeUpdateEmail,
 } from "firebase/auth";
 import { EmailAuthProvider } from "firebase/auth/web-extension";
-import { createGameProgressDocument } from "@/features/game/firestore";
+import { createGameProgressDocument } from "@/features/game/createGameProgress";
 import { auth } from "@/firebase/config";
 
 export async function signupWithEmail(email: string, password: string) {
