+++ src/features/game/firestore.tsx
@@ -1,32 +1,13 @@
-import { collection, deleteDoc, doc, getDocs, setDoc } from "firebase/firestore";
+import { collection, doc } from "firebase/firestore";
 import { db } from "@/firebase/config";
 
-export async function createGameProgressDocument(uid: string) {
-  try {
-    const objectInfoCollectionRef = collection(db, "ObjectInfo");
-    const objectInfoQuerySnap = await getDocs(objectInfoCollectionRef);
-    if(objectInfoQuerySnap.empty) {
-      return false;
-    }
-    const gameProgressDataInit: {[key: string]: boolean} = {};
-    objectInfoQuerySnap.forEach((doc) => {
-      gameProgressDataInit[doc.id] = false;
-    });
-
-    const GameProgressDocRef = doc(db, "game_progress", uid);
-    await setDoc(GameProgressDocRef, gameProgressDataInit);
-    return true;
-  } catch {
-    return false;
-  }
+export const collections = {
+  objectInfo: collection(db, "ObjectInfo"),
 }
 
-export async function deleteDocument(uid: string) {
-  try {
-    const docRef = doc(db, "game_progress", uid);
-    await deleteDoc(docRef);
-    return true;
-  } catch {
-    return false;
-  }
+export const docRefs = {
+  ObjectInfo: (docID: string) => doc(db, "ObjectInfo", docID),
+  UUIDMap: (docID: string) => doc(db, "UUIDMap", docID),
+  GameProgress: (docID: string) => doc(db, "game_progress", docID)
 }
+
