--- src/features/game/stampData.ts
@@ -1,26 +1,0 @@
-// TODO: スタンプIDをハッシュ化
-export const stampIDs = [
-  "object_1",
-  "object_2",
-  "object_3",
-  "object_4",
-  "object_5",
-  "object_6",
-  "object_7",
-  "object_8",
-  "object_9",
-  "object_10",
-];
-
-export const stampName = [
-  "入口",
-  "階段",
-  "駐車場",
-  "ベンチ",
-  "すべり台",
-  "時計",
-  "Undefined1",
-  "Undefined2",
-  "Undefined3",
-  "Undefined4",
-];
